Modified version to 0.4.0-RC1.
The reason for changing back from 0.8.0 to 0.4.0 is the realization
that a tight coupling of the version numbers for the UI and core
does not make sense if the UI is to be allowed to evolve
independently from / faster than the core. For example, if the
versioning was strictly synchronized, every new feature in the UI
which results in a new version would require an update of the core
version number, even if nothing has changed there.

Explicitly decoupling the version numbers thus allows for more
flexibility, and if the versions are different, there is also no
implicit assumption from users expecting them to be the same.

The existing release-0.8 branch, the corresponding tags and the
published artificats will be removed shortly.