bugfix: compute image slice for 2D views, such that slices is guaranteed to be behind other sliced actors.
The slicing mechanism for the 2D slice views computes the slices only at grid positions. This commmit
makes sure that the grid position is computed such that the resulting slice is guaranteed to lie behind the sliced actor (which are sliced exactly) .
This fixes a bug, which caused that triangle meshes were covered by the image slice (and hence were not visible).