Added possibility to load shape models from h5 files with model catalog
Some h5 files specify the model not in root location, but use a model catalog,
which may have several entries pointing to location in the file where the SSM is stored.
This change makes it possible to load such shape models. If several valid shape models
are present, a selection dialog is shown to the user.

Fixes #11