Update to redesigned scalismo  (#54)
This version of Scalismo-ui builds upon the Scalismo version with the new domain concept and the recent major refactorings in the transformations. The implementation needs a cleanup, but should have all the needed functionality. 

In addition to the refactoring, the Scala version was bumped to 2.13 and support for Scala 2.11 was dropped.