added scaling to LandmarkNode
Landmarks are visualized using an ellipsoid. So far, the size of this ellipsoid was chosen to reflect one standard deviation of its associated uncertainty. This commit adds the possibility to scale this ellipsoid with a fixed scale factor. This is useful when the uncertainty is very small or large in relationship to the visualized object.

The implementation adds a new ScalingProperty and ScalingPropertyPanel, which could be reused for adding this property to other nodes.