updated to scalismo 0.91-RC2 (and vtk 9.1) (#57)
Changes required due to the update to vtk version 9.1

In previous versions, the vtkJoglPanelComponent was reimplemented and
extended. This caused problems on MacOS. It seems that the bugs that
were the reason that it was reimplemented in the first place are not
causing any problems (or are fixed?) we revert to the original
implementation.