Fixes 2 bugs that were introduced transitioning to vtk9.1
In the transitioning to vtk 9.1, we accidentaly dropped the code
that was necessary for correct handling of events in the vtk windows.
As a result the mouse wheel no longer worked and the 2D planes could
be moved. This commit fixes the problem by reintroducing the
functionality that was dropped before.