Changed catalog logic in shape model loading.
When the catalog cannot be read it is now
assumed that the model starts at the root
of the file. This change was introduced as
the reader in scalismo does not return the
dedicated error message. However, it
anyway does make little sense to distinguish
this failure case, as the error is handled
correctly down the line.