correctly treat services in sbt-assembly (#60)
With the last release of the native libs, the platform binaries
are included using the service interface. It turned out that the
existing strategy of taking the first of several files in case of
duplicate files, would no longer work. The individual service
implementation for the platform would not all be there and hence
an exception would be thrown when loading the application.

This commit fixes this problem by a new rule for the service files
in the assembly.